+++ src/Light.h
@@ -15,7 +15,7 @@ struct Light {
     LightType type;
     Vector position;
     Light(){type = LT_NONE; brightness = 1; position.set(0, 0, 0);}
-    float brightness;
+    double brightness;
 };
 
 struct PointLight: public Light {
@@ -27,7 +27,7 @@ struct AreaLight: public Light {
     //Inspired by PovRay (http://www.povray.org/documentation/view/3.6.0/313/)
     //Two direction vectors and their lengths
     Vector dir1, dir2;
-    float size1, size2;
+    double size1, size2;
 
     AreaLight(){type = LT_AREA; dir1.set(1, 0, 0); dir2.set(0, 0, -1);
         size1 = 1; size2 = 1;}
